Add option to build a static librazer and use it for Debian
librazer is a private library, not meant to be used by external
software. Installing private libraries in /usr/lib/ is not
recommended by the Debian policy, so add a CMake option (disabled
by default) to build librazer as a static library and use it for the
package build, and avoid shipping it in the razercfg package.

Fixes Lintian Warnings:
  package-name-doesnt-match-sonames librazer1
  non-dev-pkg-with-shlib-symlink usr/lib/librazer.so.1 usr/lib/librazer.so

Signed-off-by: Luca Boccassi <luca.boccassi@gmail.com>