Fix the Lachesis button function definitions.
On the Lachesis, the profile and windows button definitions are swapped around
in two ways.

Profile switch up is 0x1A instead of 0xB.
Profile switch down is 0x1B instead of 0xA.
Windows button 4 is 0xA instead of 0x1B.
Windows button 5 is 0xB instead of 0x1A.

So not only are profile and windows button swapped on which has bit 5 set, but
the two buttons swap which is first in both cases as well.